Update integeration testing for listing API
